MVKInstance: Make vkGetPhysicalDeviceToolProperties() instance again.
The Vulkan spec [states][1] that `vkGetInstanceProcAddr()` returns all
dispatchable functions of a `VkInstance` or its direct children. It
further [states][2] that `vkGetDeviceProcAddr()` does _not_ return any
dispatchable function of a `VkPhysicalDevice`. This means that, despite
that `VK_EXT_tooling_info` is a device extension, the
`vkGetPhysicalDeviceToolProperties()` function is an _instance_
function. This case falls under the "available device extension
dispatchable command for `instance`" case in Table 1, and footnote 3 in
Table 2.

Fixes the `dEQP-VK.api.version_check.entry_points` test.

[1]: https://docs.vulkan.org/spec/latest/chapters/initialization.html#vkGetInstanceProcAddr-behavior
[2]: https://docs.vulkan.org/spec/latest/chapters/initialization.html#vkGetDeviceProcAddr-behavior

Signed-off-by: Chip Davis <cdavis5x@gmail.com>